requirements.txt
streamlit>=1.32.0
python-dotenv>=1.0.0
pymongo>=4.6.0
requests>=2.31.0
pydantic>=2.7.0
openai>=1.30.0
groq>=0.9.0
google-generativeai>=0.5.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
langchain>=0.2.0
langchain-community>=0.2.0